requirements.txt
speedtest-cli
GPUtil
psutil
requests
